Friday Project_10_ AI Newsletter Project/requirements.txt
openai==0.28.0
python-dotenv
requests==2.28.0  # Specify a version of requests for consistency (adjust the version if needed)
